// document-models/subscription-instance/gen/requests/actions.ts
import type { Action } from "document-model";
import type {
  CreateClientRequestInput,
  ApproveRequestInput,
  RejectRequestInput,
  WithdrawRequestInput,
} from "../types.js";

export type CreateClientRequestAction = Action & {
  type: "CREATE_CLIENT_REQUEST";
  input: CreateClientRequestInput;
};
export type ApproveRequestAction = Action & {
  type: "APPROVE_REQUEST";
  input: ApproveRequestInput;
};
export type RejectRequestAction = Action & {
  type: "REJECT_REQUEST";
  input: RejectRequestInput;
};
export type WithdrawRequestAction = Action & {
  type: "WITHDRAW_REQUEST";
  input: WithdrawRequestInput;
};

export type SubscriptionInstanceRequestsAction =
  | CreateClientRequestAction
  | ApproveRequestAction
  | RejectRequestAction
  | WithdrawRequestAction;